requirements.txt
python-dotenv 
proxmoxer
uvicorn
fastapi
requests
pyyaml
ansible
ansible-runner